# docs/swagger.yaml
components:
  schemas:
    api.authResponse:
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/api.userResponse'
      type: object
    api.authorResponse:
      properties:
        avatar_url:
          type: string
        id:
          type: string
        username:
          type: string
      type: object
...
    api.createMediaItem:
      properties:
        media_type:
          enum:
          - image
          - video
          type: string
        media_url:
          type: string
        sort_order:
          type: integer
      required:
      - media_type
      - media_url
      type: object
    api.createPostRequest:
      properties:
        content:
          type: string
        media:
          items:
            $ref: '#/components/schemas/api.createMediaItem'
          type: array
          uniqueItems: false
        title:
          maxLength: 200
          type: string
      required:
      - content
      - title
      type: object
    api.createPostResponse:
      properties:
        id:
          type: string
      type: object
...
    api.getPostsResponse:
      properties:
        author:
          $ref: '#/components/schemas/api.authorResponse'
        content:
          type: string
        created_at:
          type: string
        id:
          type: string
        media:
          items:
            $ref: '#/components/schemas/api.mediaResponse'
          type: array
          uniqueItems: false
        title:
          type: string
        view_count:
          type: integer
      type: object
    api.listPostsItemResponse:
      properties:
        author:
          $ref: '#/components/schemas/api.authorResponse'
        content:
          type: string
        cover_url:
          type: string
        created_at:
          type: string
        id:
          type: string
        title:
          type: string
        view_count:
          type: integer
      type: object
    api.loginRequest:
      properties:
        password:
          type: string
        username:
          type: string
      required:
      - password
      - username
      type: object
    api.mediaResponse:
      properties:
        id:
          type: string
        media_type:
          type: string
        media_url:
          type: string
        sort_order:
          type: integer
      type: object
    api.registerRequest:
      properties:
        password:
          maxLength: 128
          minLength: 6
          type: string
        username:
          maxLength: 32
          minLength: 3
          type: string
      required:
      - password
      - username
      type: object
...
    api.updateProfileRequest:
      properties:
        avatar_url:
          type: string
        bio:
          maxLength: 200
          type: string
        username:
          maxLength: 32
          minLength: 3
          type: string
      required:
      - username
      type: object
    api.userResponse:
      properties:
        avatar_url:
          type: string
        bio:
          type: string
        id:
          type: string
        username:
          type: string
      type: object
    render.Response-api_authResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.authResponse'
        msg:
          type: string
      type: object
...
    render.Response-api_createPostResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.createPostResponse'
        msg:
          type: string
      type: object
...
    render.Response-api_getPostsResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.getPostsResponse'
        msg:
          type: string
      type: object
...
    render.Response-api_userResponse:
      properties:
        code:
          type: integer
        data:
          $ref: '#/components/schemas/api.userResponse'
        msg:
          type: string
      type: object
...
    render.Response-array_api_listPostsItemResponse:
      properties:
        code:
          type: integer
        data:
          items:
            $ref: '#/components/schemas/api.listPostsItemResponse'
          type: array
          uniqueItems: false
        msg:
          type: string
      type: object
    render.ResponseWithoutData:
      properties:
        code:
          type: integer
        msg:
          type: string
      type: object
    render.errorResponse:
      properties:
        code:
          type: integer
        msg:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      scheme: bearer
      type: http
...
info:
  description: 小蓝书后端接口文档
  title: 小蓝书 API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /posts:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: 每页数量
        in: query
        name: page_size
        schema:
          default: 20
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-array_api_listPostsItemResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      summary: 获取帖子列表
      tags:
      - posts
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.createPostRequest'
              description: 帖子内容
              summary: body
        description: 帖子内容
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_createPostResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 创建帖子
      tags:
      - posts
  /posts/{id}:
    delete:
      parameters:
      - description: 帖子 ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.ResponseWithoutData'
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 删除帖子
      tags:
      - posts
    get:
      parameters:
      - description: 帖子 ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_getPostsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Not Found
      summary: 获取帖子详情
      tags:
      - posts
...
  /posts/user/{userID}:
    get:
      parameters:
      - description: 用户 ID
        in: path
        name: userID
        required: true
        schema:
          type: string
      - description: 页码
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: 每页数量
        in: query
        name: page_size
        schema:
          default: 20
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-array_api_listPostsItemResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      summary: 获取指定用户的帖子列表
      tags:
      - posts
  /users/{id}:
    get:
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_userResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Not Found
      summary: 获取用户资料
      tags:
      - users
...
  /users/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.loginRequest'
              description: 登录信息
              summary: body
        description: 登录信息
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_authResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      summary: 用户登录
      tags:
      - users
  /users/profile:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.updateProfileRequest'
              description: 更新信息
              summary: body
        description: 更新信息
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_userResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 更新用户资料
      tags:
      - users
  /users/register:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.registerRequest'
              description: 注册信息
              summary: body
        description: 注册信息
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.Response-api_authResponse'
          description: OK
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/render.errorResponse'
          description: Internal Server Error
      summary: 用户注册
      tags:
      - users
servers:
- description: 本地开发环境
  url: http://localhost:8080/api/v1